+++ mcp-server/lib/server/internal/handlers/resources-read.js
@@ -1,12 +1,22 @@
 import { getBy } from '../../external/kv.js';
 import { InternalServerError } from './error.js';
 
-export function createResourcesReadHandler(logger, mcpServers) {
+export function createResourcesReadHandler(
+  logger,
+  mcpServers,
+  requestedMcpServer
+) {
   return async req => {
+    const path = req.path;
     const uri = req.params.uri;
-    logger.debug('Resource read', uri);
+    logger.debug('Resource read', path, uri);
 
-    const readResource = getBy(mcpServers, uri, 'readResource');
+    const readResource = getBy(
+      mcpServers,
+      requestedMcpServer,
+      uri,
+      'readResource'
+    );
     if (!readResource) {
       throw new InternalServerError(`Resource not found: ${uri}`);
     }